[Fix 🪛] API 속도 향상
