--- src/components/NotificationModal/apis/seenNotificationAPI.ts
@@ -1,11 +1,0 @@
-import { AUTH_API } from "@/apis/Api"
-
-const seenNotificationAPI = async () => {
-  await AUTH_API.put("/notifications/seen")
-    .then()
-    .catch(() => {
-      throw new Error("알림 목록 받아 오기 실패")
-    })
-}
-
-export default seenNotificationAPI
