+++ src/components/NotificationModal/hooks/useSeenNotification.ts
@@ -1,5 +1,6 @@
 import { useMutation, useQueryClient } from "@tanstack/react-query"
-import seenNotificationAPI from "../apis/seenNotificationAPI"
+import { AUTH_API } from "@/apis/Api"
+
 import { QUERY_KEY_GET_NOTIFICATION } from "./useGetNotification"
 const QUERY_KEY_SEEN_NOTIFICATION = "SEEN_NOTIFICATION"
 const useSeenNotification = () => {
@@ -18,3 +19,11 @@ const useSeenNotification = () => {
 }
 
 export default useSeenNotification
+
+const seenNotificationAPI = async () => {
+  await AUTH_API.put("/notifications/seen")
+    .then()
+    .catch(() => {
+      throw new Error("알림 목록 받아 오기 실패")
+    })
+}
