+++ src/components/PostDetail/hooks/useUnLikePost.tsx
@@ -4,6 +4,7 @@ import useModal from "@/components/Modal/hooks/useModal"
 import { POST_DETAIL_ERROR_MESSAGE } from "@/constants/errorMessage"
 import useAuthUserStore from "@/stores/useAuthUserStore"
 import { useMutation, useQueryClient } from "@tanstack/react-query"
+import { POST_DETAIL_QUERY_KEY } from "./useGetPost"
 
 const MUTATION_KEY_UN_LIKE_POST_KEY = "IT_IS_UN_LIKE_MUTATION_KEY_5448718927139"
 
@@ -16,7 +17,9 @@ const useUnLikePost = () => {
     mutationKey: [MUTATION_KEY_UN_LIKE_POST_KEY],
     mutationFn: fetchUnLikePost,
     onSuccess: () => {
-      queryClient.refetchQueries()
+      queryClient.invalidateQueries({
+        queryKey: [POST_DETAIL_QUERY_KEY],
+      })
       AUTH_API.get("/auth-user")
         .then((res) => res.data)
         .then((userData) => updateUser(userData))
